add entering call lbl in case of break active
