+++ intellectus-desktop-client/Views/Suggestions/OnCallWindow.cs
@@ -64,7 +64,7 @@ namespace intellectus_desktop_client
 
         private void startCall()
         {
-            if (CanStartCall())
+            if (Services.CallService.CanStartCall())
             {
                 if (API.StartCall())
                 {
@@ -74,22 +74,6 @@ namespace intellectus_desktop_client
             }
             else
                 MessageBox.Show("No se puede iniciar una llamada, se encuentra en un descanso");
-        }
-
-        private bool CanStartCall()
-        {
-            var callAndBreak = Domain.CurrentUser.Call == null || Domain.CurrentUser.Call.BreakAssigned == false;
-            if (!callAndBreak)
-            {
-                var endTime = Domain.CurrentUser.Call.EndTime;
-                var minutes = Domain.CurrentUser.Call.MinutesDuration;
-                var breakEndTime = endTime.AddMinutes(minutes);
-                return DateTime.UtcNow > breakEndTime;
-            }
-            else
-                return true;
-            
-                
         }
     }
 }
